backend/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.33.0
python-dotenv==1.0.1
pydantic==2.10.6
TikTokApi==7.1.0
playwright==1.53.0
pytest-asyncio==1.1.0